feat: separate 'list'
